# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Katsumori API documentation",
        "contact": {
            "email": "administrator@katsumori.fun"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:80/api/v1",
            "description": "v1"
        }
    ],
    "paths": {
...
        "/devices": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Отображение списка устройств, с которых был выполнен вход в учетную запись",
                "description": "Отображение всех устройств, с которых был выполнен вход в учетную запись",
                "operationId": "abc5c12998275c2baab51c06b84fe311",
                "responses": {
                    "200": {
                        "description": "Cписка устройств",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Device"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка аутентификации (просроченный токен, либо его отсутствие)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/devices/current": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Отображение информаци об устройстве, с которого была произведена авторизация",
                "description": "Отображение информаци об устройстве, с которого была произведена авторизация",
                "operationId": "51409af7f668b0bc9a9739298f1c0cde",
                "responses": {
                    "200": {
                        "description": "Информация об устройстве",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Device"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка аутентификации (просроченный токен, либо его отсутствие)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/devices/logout/{id}": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Выход с устройства по id авторизации, который можно узнать в списке /devices/",
                "description": "Выход с устройства по id авторизации, который можно узнать в списке /devices/",
                "operationId": "3cb264d349dfd2caf2a8e9a056389f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id логина",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка аутентификации (просроченный токен, либо его отсутствие)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Id логина не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "error": {
                                                        "type": "string",
                                                        "example": "Login id not found"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/devices/logout/all": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Выход со всех устройств",
                "description": "Выход со всех устройств",
                "operationId": "04931e5d242c4bad8790ee6a11073924",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка аутентификации (просроченный токен, либо его отсутствие)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Device": {
                "title": "Device",
                "description": "Модаль устройства, с которого был выполнен вход",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Id авторизации",
                        "type": "string",
                        "format": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Дата входа",
                        "type": "string",
                        "format": "string",
                        "example": "2022-05-28T17:07:59.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Дата обновления входа",
                        "type": "string",
                        "format": "string",
                        "example": "2022-05-28T17:07:59.000000Z"
                    },
                    "user_agent": {
                        "title": "user_agent",
                        "description": "User agent",
                        "type": "string",
                        "format": "string",
                        "example": "YukiDubFunWeb"
                    },
                    "ip": {
                        "title": "ip",
                        "description": "Ip пользователя",
                        "type": "string",
                        "format": "string",
                        "example": "127.0.0.1"
                    },
                    "ip_data": {
                        "title": "ip_data",
                        "description": "Допллнительная информация (на данный момент всегда пустая)",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "device_type": {
                        "title": "device_type",
                        "description": "Тип устройства",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "device": {
                        "title": "device",
                        "description": "Устройство",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "platform": {
                        "title": "platform",
                        "description": "Платформа",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "browser": {
                        "title": "browser",
                        "description": "Браузер (если вход был произведен через него)",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "city": {
                        "title": "city",
                        "description": "Город",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "region": {
                        "title": "region",
                        "description": "Регион",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "country": {
                        "title": "country",
                        "description": "Страна",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    },
                    "is_current": {
                        "title": "is_current",
                        "description": "Это устройство",
                        "type": "string",
                        "format": "string",
                        "example": "null"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}